requirements.txt
tensorflow==1.5.0
keras==2.1.3
numpy==1.13.3
matplotlib
h5py
pillow
scikit-learn
click
future
opencv-python
tinydb
scipy
pandas
tqdm
requests
pyflann
xxhash
